% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kelly_proportion}
\alias{kelly_proportion}
\title{Kelly Criterion Proportion Calculator (Vectorized)}
\usage{
kelly_proportion(prob_win, odds)
}
\arguments{
\item{prob_win}{Numeric vector. The probability of winning the bet (between 0 and 1).}

\item{odds}{Numeric vector. The decimal odds offered by the bookmaker (must be > 0).}
}
\value{
Numeric vector. The recommended fraction of the bankroll to wager.
If the result is <= 0, no bet should be placed.
}
\description{
Calculates the optimal proportion of bankroll to wager using the Kelly criterion.
}
\examples{
kelly_proportion(c(0.6, 0.55), c(2.5, 1.91))

}
